User's can only access projects they own

+++ application/models/project_model.php
@@ -5,28 +5,28 @@
  */
 class Project_model extends CI_Model
 {
-    public function get_project($id)
-    {
-        return $this->db->get_where('projects', ['id' => $id])->row();
-    }
-
-    public function get_projects()
-    {
-        return $this->db->get('projects')->result();
-    }
-
     public function create_project($data)
     {
         return $this->db->insert('projects', $data);
     }
 
-    public function update_project($id, $data)
+    public function get_projects($user_id)
     {
-        return $this->db->where('id', $id)->update('projects', $data);
+        return $this->db->get_where('projects', ['user_id' => $user_id])->result();
     }
 
-    public function delete_project($id)
+    public function get_project($id, $user_id)
     {
-        return $this->db->where('id', $id)->delete('projects');
+        return $this->db->get_where('projects', ['id' => $id, 'user_id' => $user_id])->row();
+    }
+
+    public function update_project($id, $data, $user_id)
+    {
+        return $this->db->where(['id' => $id, 'user_id' => $user_id])->update('projects', $data);
+    }
+
+    public function delete_project($id, $user_id)
+    {
+        return $this->db->where(['id' => $id, 'user_id' => $user_id])->delete('projects');
     }
 }
